+++ src/main/java/me/despical/kotl/commands/game/LeaderBoardCommand.java
@@ -16,6 +16,7 @@ import me.despical.kotl.ConfigPreferences;
 import me.despical.kotl.api.StatsStorage;
 import me.despical.kotl.commands.SubCommand;
 import me.despical.kotl.commands.exception.CommandException;
+import me.despical.kotl.user.data.MysqlManager;
 
 /**
  * @author Despical
@@ -45,7 +46,7 @@ public class LeaderBoardCommand extends SubCommand {
 			return;
 		}
 		try {
-			StatsStorage.StatisticType statisticType = StatsStorage.StatisticType.valueOf(args[0].toUpperCase());
+			StatsStorage.StatisticType statisticType = StatsStorage.StatisticType.valueOf(args[0].toUpperCase(java.util.Locale.ENGLISH));
 			printLeaderboard(sender, statisticType);
 		} catch (IllegalArgumentException exception) {
 			sender.sendMessage(getPlugin().getChatManager().getPrefix() + getPlugin().getChatManager().colorMessage("Commands.Statistics.Invalid-Name"));
@@ -55,7 +56,7 @@ public class LeaderBoardCommand extends SubCommand {
 	private void printLeaderboard(CommandSender sender, StatsStorage.StatisticType statisticType) {
 		LinkedHashMap<UUID, Integer> stats = (LinkedHashMap<UUID, Integer>) StatsStorage.getStats(statisticType);
 		sender.sendMessage(getPlugin().getChatManager().colorMessage("Commands.Statistics.Header"));
-		String statistic = StringUtils.capitalize(statisticType.toString().toLowerCase().replace("_", " "));
+		String statistic = StringUtils.capitalize(statisticType.toString().toLowerCase(java.util.Locale.ENGLISH).replace("_", " "));
 		for (int i = 0; i < 10; i++) {
 			try {
 				UUID current = (UUID) stats.keySet().toArray()[stats.keySet().toArray().length - 1];
@@ -68,7 +69,7 @@ public class LeaderBoardCommand extends SubCommand {
 				if (this.getPlugin().getConfigPreferences().getOption(ConfigPreferences.Option.DATABASE_ENABLED)) {
 					try (Connection connection = this.getPlugin().getMysqlDatabase().getConnection()) {
 						Statement statement = connection.createStatement();
-						ResultSet set = statement.executeQuery("SELECT name FROM playerstats WHERE UUID='" + current.toString() + "'");
+						ResultSet set = statement.executeQuery("SELECT name FROM " + ((MysqlManager) getPlugin().getUserManager().getDatabase()).getTableName() + " WHERE UUID='" + current.toString() + "'");
 						if (set.next()) {
 							sender.sendMessage(formatMessage(statistic, set.getString(1), i + 1, stats.get(current)));
 							continue;